Provide global cache time support

+++ src/main/java/com/github/blindpirate/gogradle/util/DateUtils.java
@@ -1,7 +1,7 @@
 package com.github.blindpirate.gogradle.util;
 
 public class DateUtils {
-    public static long toMilliseconds(int second) {
+    public static long toMilliseconds(long second) {
         return 1000L * second;
     }
 }
